39 moved components to own folders and fixed TS errors and some styling changes
