+++ src/root.tsx
@@ -1,7 +1,7 @@
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter, Route, Routes } from 'react-router';
 import { ErrorRoute } from './routes/errorroute/errorroute.tsx';
-import { ModifyRoute } from './routes/modifyroute/modifyroute.tsx';
+import { ModifyRoute } from './routes/modifyroute/modifyroute/modifyroute.tsx';
 import { DashboardRoute } from './routes/dashboardroute/dashboardroute.tsx';
 import '../stories/_nongula.scss';
