+++ src/routes/dashboardroute/dashboardroute.scss
@@ -2,8 +2,8 @@
     display: flex;
     flex-direction: column;
     align-items: center;
-
     height: 95vh;
+
     .dashboard__header {
         display: flex;
         flex-direction: row;
@@ -17,5 +17,10 @@
         justify-content: center;
         align-items: center;
         gap: 2rem;
+    }
+    .custom-modal__actions {
+        display: flex;
+        flex-direction: row;
+        gap: 1rem;
     }
 }
