--- src/routes/modifyroute/customcaloriesmodal.scss
@@ -1,17 +1,0 @@
-.custom-modal {
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin-top: 280px;
-}
-
-.custom-modal-wrapper {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-direction: column;
-    gap: 20px;
-    min-width: 250px;
-    width: 100%;
-}
