+++ src/routes/modifyroute/customcaloriesmodal/customcaloriesmodal.scss
@@ -1,0 +1,24 @@
+.custom-modal {
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    min-width: 250px;
+    padding: 0 1rem;
+    gap: 2rem;
+    width: 100%;
+    height: 60vh;
+
+    .custom-modal__content {
+        width: 100%;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        gap: 1rem;
+    }
+    .custom-modal__actions {
+        display: flex;
+        flex-direction: column;
+        width: 100%;
+        gap: 1rem;
+    }
+}
