+++ stories/_nongula.scss
@@ -11,4 +11,4 @@
 @import './components/List/list.scss';
 @import './components/ProgressBar/progressbar.scss';
 @import './components/SegmentedControl/segmentedcontrol.scss';
-@import '../src/routes/modifyroute/customcaloriesmodal.scss';
+@import '../src/routes/modifyroute/customcaloriesmodal/customcaloriesmodal';
