+++ stories/components/Button/button.scss
@@ -20,6 +20,11 @@
     color: white;
 }
 
+.button--danger {
+    background-color: #b65050;
+    color: white;
+}
+
 .button--small {
     padding: 10px 16px;
     font-size: 0.75rem;
