+++ stories/components/TextField/textfield.scss
@@ -20,7 +20,6 @@
 }
 
 .input {
-    cursor: pointer;
     display: flex;
     border: 2px solid black;
     border-radius: 16px;
